Restructure project layout by moving `Header`, `Footer`, and `MainLayout` components to `modules/shared/` directory
