+++ src/AppRouter.tsx
@@ -1,5 +1,5 @@
 import { Route, Routes } from 'react-router-dom';
-import { MainLayout } from './layout/MainLayout';
+import { MainLayout } from './modules/shared/MainLayout';
 
 export const AppRouter = () => {
   return (
