+++ src/modules/shared/MainLayout/MainLayout.tsx
@@ -4,7 +4,6 @@ import { Footer } from '../Footer';
 
 export const MainLayout = () => {
   return (
-    <>
       <Header />
       <main>
         <Outlet />
